server.c
Usage: server [-p port] /path/to/root

Note that port is optional - if not entered this will default to 8080. 

In brief, server.c behaves in the following way:

The server is initiated with the function "start" which takes the port number and file path to the server's root as arguments (line 114). An infinite loop then begins on line 131, which ensures that the server continually listens on this port for any incoming HTTP requests until the user terminates the program with CTRL-C (calling the function "stop" on line 158). Only one connection to the server is accepted at any one time, which is tested for by the function "connected" (line 162), with this returning true if so. The function "request" (line 165) reads directly from the currently open socket (or client file descriptor, cfd) the HTTP request from the client, which is stored in memory as char* message. The length of the message is also recorded in memory as size_t length. Lines 167-178 strip the HTTP request down to the template: 

method SP request-target SP HTTP-version CRLF.

Line 186 uses the function "parse" to extract the abs_path and any query associated with the request and store these in memory, while returning true if the HTTP request is deemed valid. In turn, on line 197-205, the absolute path is concatenated with the server's root, to provide a local path to the requested file. If this file path is valid (line 208), the program then checks if this is a path to a directory or an actual file. If a directory, the function "indexes" (line 229) is used to check whether or not the file type index.html or index.php are present within. If so, the variable path is reassigned to this location. If not, the function "list" (line 239) merely lists the content of the directory. In contrast, if the user has requested an actual file, the function "lookup" first looks-up the file-type requested ie .html, .php etc. If the file is of type php, then the function "interpret" (line 255) is employed to display this dynamic content. If any other file type is requested, the function "transfer" (line 261) is employed. If none of the above cases apply, an error of 501 is displayed. 

The majority of the code and functionality shown above was written by CS50. However, to ensure that the program works, the student (myself in this case) was required to implement the following functions: 

lookup (implemented on line 764) - returns file extension of file associated with local path
parse (implemented on line 829) - Parses HTTP request, stored in variable "line". Ensures request is valid, while extracting and storing in memory the absolute path of the file requested as well as any query associated with the request, ie following "q".
load (implemented on line 702) - Called by "interpret" and "transfer" to load content of php interpreter's output into memory at char* content, which is then used by the function "respond" at line 598 to return the content to the user. 
indexes(implemented on line 446) - Determines whether there is a file of type index.html or index.php inside of the directory requested by the client and if so return as new path to that file.